+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/buf/project/BufFileManager.kt
@@ -3,15 +3,20 @@ package io.kanro.idea.plugin.protobuf.buf.project
 import com.fasterxml.jackson.core.JacksonException
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.dataformat.yaml.YAMLMapper
+import com.intellij.ide.DataManager
+import com.intellij.ide.DefaultTreeExpander
+import com.intellij.openapi.actionSystem.ActionGroup
+import com.intellij.openapi.actionSystem.ActionManager
+import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.application.ReadAction
 import com.intellij.openapi.components.BaseState
 import com.intellij.openapi.components.PersistentStateComponent
 import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
-import com.intellij.openapi.diagnostic.Logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ex.ProjectRootManagerEx
+import com.intellij.openapi.ui.SimpleToolWindowPanel
 import com.intellij.openapi.util.EmptyRunnable
 import com.intellij.openapi.vfs.VfsUtil
 import com.intellij.openapi.vfs.VirtualFile
@@ -26,9 +31,10 @@ import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.psi.PsiElement
 import com.intellij.psi.search.FilenameIndex
 import com.intellij.psi.search.ProjectScope
+import com.intellij.psi.util.CachedValueProvider
+import com.intellij.psi.util.CachedValuesManager
 import com.intellij.ui.ScrollPaneFactory
 import com.intellij.ui.content.ContentFactory
-import com.intellij.ui.treeStructure.Tree
 import com.intellij.util.concurrency.NonUrgentExecutor
 import com.intellij.util.xmlb.annotations.Tag
 import com.intellij.util.xmlb.annotations.XCollection
@@ -41,18 +47,17 @@ import io.kanro.idea.plugin.protobuf.buf.schema.common.BufLockDepsRemoteSchema
 import io.kanro.idea.plugin.protobuf.buf.schema.common.BufLockDepsRepositorySchema
 import io.kanro.idea.plugin.protobuf.buf.schema.common.BufNameFieldSchema
 import io.kanro.idea.plugin.protobuf.buf.schema.v1.BufWorkDirectoriesFieldSchema
-import io.kanro.idea.plugin.protobuf.buf.ui.BufToolWindowRootElement
+import io.kanro.idea.plugin.protobuf.buf.ui.toolwindow.BufToolWindowRootElement
 import io.kanro.idea.plugin.protobuf.buf.util.BUF_LOCK
 import io.kanro.idea.plugin.protobuf.buf.util.BUF_WORK_YAML
 import io.kanro.idea.plugin.protobuf.buf.util.BUF_YAML
 import io.kanro.idea.plugin.protobuf.buf.util.BufFiles
-import io.kanro.idea.plugin.protobuf.ui.SmartTreeCellRenderer
+import io.kanro.idea.plugin.protobuf.ui.SmartTree
 import io.kanro.idea.plugin.protobuf.ui.SmartTreeModel
 import java.util.Stack
 
 @State(name = "BufFileManager", storages = [Storage("protobuf.xml")])
 class BufFileManager(val project: Project) : PersistentStateComponent<BufFileManager.State> {
-    private val logger = Logger.getInstance(BufFileListener::class.java)
     private val state = State()
     private val yamlMapper = YAMLMapper()
     private val cacheRootPointer = VirtualFilePointerManager.getInstance().createDirectoryPointer(
@@ -67,6 +72,24 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
             }
         }
     )
+    private val libraryLookup: Map<String, State.Module>
+        get() = CachedValuesManager.getManager(project).getCachedValue(project) {
+            val result = state.libraries.associateBy { it.reference ?: "unknown" }
+            CachedValueProvider.Result.create(result, state)
+        }
+    private val moduleLookup: Map<String, State.Module>
+        get() = CachedValuesManager.getManager(project).getCachedValue(project) {
+            val result = state.modules.associateBy { it.path ?: "unknown" }
+            CachedValueProvider.Result.create(result, state)
+        }
+    private val workspaceLookup: Map<String, State.Workspace>
+        get() = CachedValuesManager.getManager(project).getCachedValue(project) {
+            val result = state.workspaces.associateBy { it.path ?: "unknown" }
+            CachedValueProvider.Result.create(result, state)
+        }
+    private val treeModel by lazy {
+        SmartTreeModel(BufToolWindowRootElement(this))
+    }
 
     init {
         importProject()
@@ -97,6 +120,8 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
             old.copyFrom(new)
             state.modules.sortBy { it.path }
         }
+
+        updateToolWindow(state.modules.isNotEmpty())
     }
 
     fun workspaceChanged(oldRoot: VirtualFile?, newRoot: VirtualFile?) {
@@ -154,7 +179,7 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
         }
     }
 
-    private fun importProject() {
+    fun importProject() {
         ReadAction.nonBlocking {
             if (this.project.isDisposed) return@nonBlocking
             state.modules.clear()
@@ -264,30 +289,64 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
         }
     }
 
-    fun registerToolWindow(show: Boolean) {
+    private fun registerToolWindow(show: Boolean) {
         if (!show) return
         ApplicationManager.getApplication().invokeLater {
             val manager = ToolWindowManager.getInstance(project)
-            val toolWindow = manager.registerToolWindow(
-                RegisterToolWindowTask.notClosable(
-                    "buf", Icons.BUF_LOGO, ToolWindowAnchor.RIGHT
+            val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(TOOL_WINDOW_ID)
+            if (toolWindow == null) {
+                createToolWindowContent(
+                    manager.registerToolWindow(
+                        RegisterToolWindowTask.notClosable(
+                            TOOL_WINDOW_ID, Icons.BUF_LOGO, ToolWindowAnchor.RIGHT
+                        )
+                    )
                 )
-            )
-            createToolWindowContent(toolWindow)
+            } else {
+                updateToolWindow(show)
+            }
         }
     }
 
-    fun createToolWindowContent(toolWindow: ToolWindow) {
-        val model = SmartTreeModel(BufToolWindowRootElement(project))
-        val tree = Tree(model).apply {
-            this.isRootVisible = false
-            this.cellRenderer = SmartTreeCellRenderer()
+    private fun createToolWindowContent(toolWindow: ToolWindow) {
+        toolWindow.stripeTitle = "Buf"
+        toolWindow.title = "Buf"
+        treeModel.reload()
+        val tree = SmartTree(treeModel)
+        val panel = SimpleToolWindowPanel(true)
+        panel.toolbar = ActionManager.getInstance().createActionToolbar(
+            "Protobuf.Buf",
+            ActionManager.getInstance().getAction("io.kanro.idea.plugin.Protobuf.Buf.ToolWindow") as ActionGroup,
+            true
+        ).let {
+            it.targetComponent = tree
+            it.component
         }
-        val panel = ScrollPaneFactory.createScrollPane(tree)
+        panel.setContent(ScrollPaneFactory.createScrollPane(tree, true))
+        DataManager.registerDataProvider(panel) {
+            if (it == PlatformDataKeys.TREE_EXPANDER.name) {
+                DefaultTreeExpander(tree)
+            } else null
+        }
         val content = ContentFactory.SERVICE.getInstance().createContent(
             panel, null, false
         )
         toolWindow.contentManager.addContent(content)
+    }
+
+    private fun updateToolWindow(show: Boolean) {
+        ApplicationManager.getApplication().invokeLater {
+            val toolWindow = ToolWindowManager.getInstance(project).getToolWindow(TOOL_WINDOW_ID)
+                ?: return@invokeLater registerToolWindow(show)
+
+            if (!show) {
+                toolWindow.hide()
+                toolWindow.isShowStripeButton = false
+                return@invokeLater
+            }
+            treeModel.reload()
+            toolWindow.show()
+        }
     }
 
     override fun getState(): State {
@@ -300,6 +359,15 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
         if (state.modificationCount != modificationCount) {
             notifyLibraryRootsChanged()
         }
+    }
+
+    fun findLibrary(name: String): State.Module? {
+        return libraryLookup[name]
+    }
+
+    fun findLibrary(dep: State.Dependency?): State.Module? {
+        dep ?: return null
+        return findLibrary(dep.name())
     }
 
     fun findModule(file: VirtualFile): State.Module? {
@@ -340,7 +408,7 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
     }
 
     fun resolveDependencies(deps: Collection<State.Dependency>): List<State.Module> {
-        val libraries = state.libraries.associateBy { it.reference }
+        val libraries = libraryLookup
         val required = Stack<State.Dependency>()
         val result = mutableListOf<State.Module>()
         val resolved = mutableSetOf<String>()
@@ -431,5 +499,9 @@ class BufFileManager(val project: Project) : PersistentStateComponent<BufFileMan
                 return name()
             }
         }
+    }
+
+    companion object {
+        val TOOL_WINDOW_ID = "buf"
     }
 }