+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/ui/TooltipPresentation.kt
@@ -1,5 +1,5 @@
 package io.kanro.idea.plugin.protobuf.ui
 
 interface TooltipPresentation {
-    fun tooltip(): String
+    fun tooltip(): String?
 }