+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/ui/TreeElement.kt
@@ -3,3 +3,7 @@ package io.kanro.idea.plugin.protobuf.ui
 interface TreeElement {
     fun children(): Array<TreeElement>
 }
+
+interface TreeActionElement : TreeElement {
+    fun doubleClickOrEnter()
+}